fix(apps/ekiosk): additional warning fixes
- Multiple files: Various warning fixes across the codebase

+++ apps/EKiosk/src/Services/CashDispenserManager.cpp
@@ -227,7 +227,7 @@ void CashDispenserManager::onUnitsDefined() {
 
         if (units != 0) {
             int currentUnits =
-                m_CurrencyCashList.contains(configPath) ? m_CurrencyCashList[configPath].size() : 0;
+                m_CurrencyCashList.contains(configPath) ? static_cast<int>(m_CurrencyCashList[configPath].size()) : 0;
 
             if (currentUnits < units) {
                 PPSDK::TCashUnitList cashUnitList = PPSDK::TCashUnitList(
